% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlindBuffer.R
\name{BlindBuffer}
\alias{BlindBuffer}
\title{BlindBuffer}
\usage{
BlindBuffer(Line, BuffDist, EraseDist)
}
\arguments{
\item{Line}{An object of class SpatVector that must be a line.}

\item{BuffDist}{The distance to be buffered.}

\item{startpt}{The distance for the blind spot buffer.}
}
\description{
Generates a buffer for each line segment with the blind spot near the line removed.
}
\examples{
BlindBuffer(Line, 685, 22)
}
