% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineReverse.R
\name{LineReverse}
\alias{LineReverse}
\title{LineReverse}
\usage{
LineReverse(Line)
}
\arguments{
\item{Line}{An object of class SpatVector that must be a line.}
}
\description{
Switches the direction of a SpatVector line
}
\examples{
LineReverse(Line)
}
