% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitLine.R
\name{splitLine}
\alias{splitLine}
\title{LineReverse}
\usage{
splitLine(Line, interval = 6, startpt = 1)
}
\arguments{
\item{Line}{An object of class SpatVector that must be a line.}

\item{interval}{An integer, the number of points to be included in each line segment.}

\item{startpt}{An integer, the position of the first split. Used to offset the buffers, a line from 1 to this number will be created.}
}
\description{
Splits a line into multiple segments by a pulling a number of points.
}
\examples{
splitLine(Line, interval = 6, startpt = 1)
}
