% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewDirCheck.R
\name{ViewDirCheck}
\alias{ViewDirCheck}
\title{ViewDirCheck}
\usage{
ViewDirCheck(Lines, DEM, Outer = 200, Inner = 100)
}
\arguments{
\item{DEM}{An object of class SpatRaster. Elevation data.}

\item{Outer}{A numeric value for the distance of the outer buffer of the search area.}

\item{Inner}{A numeric value for the distance of the inner erase buffer of the search area.}

\item{Line}{An object of class SpatVector that must be a line.}
}
\description{
Checks the specified view direction of tracklogs against a DEM. Flags mismatches.
}
\examples{
ViewDirCheck(Lines, DEM)

}
